Create 1.Write a C program to find area and circumference of circle.
In this tutorial, we will learn how to write a C++ program to calculate the area and circumference of a circle, given its radius. We will start by defining the necessary variables and constants, and then we will prompt the user to enter the radius value using the standard input stream (cin). We will use the mathematical formulae for the area (pir^2) and circumference (2pi*r) to compute the results, where pi is a constant value defined as 3.14159. Finally, we will output the calculated values to the standard output stream (cout), formatted to two decimal places. We will also add some error checking to handle invalid input values. This program will be helpful for anyone interested in learning how to perform basic mathematical calculations in C++