Create C program to read three numbers and find largest of three
n this tutorial, we will learn how to write a C program to find the largest of three numbers entered by the user. We will start by defining the necessary variables and constants, and then we will prompt the user to enter the three values using the standard input stream (scanf). We will use conditional statements (if-else) to compare the three numbers and determine which one is the largest. We will then output the result to the standard output stream (printf), indicating which number is the largest. We will also add some error checking to handle invalid input values. This program will be useful for anyone interested in learning how to perform basic comparison operations in C.